pyproject.toml
[project]
name = "lotradioalerts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "boto3>=1.34.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "selenium>=4.27.1",
    "spotipy>=2.25.0",
]
